chore: pedantic and nursery lints stand guard in the manifest, the easy debts paid

+++ crates/hydebar-gui/src/app/update/outputs.rs
@@ -69,7 +69,7 @@ impl App {
 
         let surface_scale = self.config.appearance.scale_factor as f32;
 
-        self.screen_height = Some(geometry.pixels.1 as f32 / geometry.scale.max(f32::EPSILON));
+        self.screen_height = Some(geometry.pixels.1 / geometry.scale.max(f32::EPSILON));
 
         let metrics = auto_metrics(
             geometry.pixels.0,
